Rename and move existing Java remote debugging code
