+++ package.json
@@ -48,7 +48,7 @@
         "onCommand:azureFunctions.deleteFunction",
         "onCommand:azureFunctions.deploy",
         "onCommand:azureFunctions.configureDeploymentSource",
-        "onCommand:azureFunctions.debugFunctionAppOnAzure",
+        "onCommand:azureFunctions.startJavaRemoteDebug",
         "onCommand:azureFunctions.appSettings.add",
         "onCommand:azureFunctions.appSettings.download",
         "onCommand:azureFunctions.appSettings.upload",
@@ -176,8 +176,8 @@
                 "category": "Azure Functions"
             },
             {
-                "command": "azureFunctions.debugFunctionAppOnAzure",
-                "title": "%azFunc.debugFunctionAppOnAzure%",
+                "command": "azureFunctions.startJavaRemoteDebug",
+                "title": "%azFunc.startJavaRemoteDebug%",
                 "category": "Azure Functions"
             },
             {
@@ -452,8 +452,8 @@
                     "group": "4@2"
                 },
                 {
-                    "command": "azureFunctions.debugFunctionAppOnAzure",
-                    "when": "view == azFuncTree && viewItem =~ /^azFunc(Production|)Slot$/ && config.azureFunctions.enableRemoteDebugging == true",
+                    "command": "azureFunctions.startJavaRemoteDebug",
+                    "when": "view == azFuncTree && viewItem =~ /^azFunc(Production|)Slot$/ && config.azureFunctions.enableJavaRemoteDebugging == true",
                     "group": "5@1"
                 },
                 {
@@ -646,8 +646,8 @@
                     "when": "never"
                 },
                 {
-                    "command": "azureFunctions.debugFunctionAppOnAzure",
-                    "when": "config.azureFunctions.enableRemoteDebugging == true"
+                    "command": "azureFunctions.startJavaRemoteDebug",
+                    "when": "config.azureFunctions.enableJavaRemoteDebugging == true"
                 },
                 {
                     "command": "azureFunctions.createSlot",
@@ -826,9 +826,9 @@
                         "description": "%azFunc.show64BitWarningDescription%",
                         "default": true
                     },
-                    "azureFunctions.enableRemoteDebugging": {
+                    "azureFunctions.enableJavaRemoteDebugging": {
                         "type": "boolean",
-                        "description": "%azFunc.enableRemoteDebugging%",
+                        "description": "%azFunc.enableJavaRemoteDebugging%",
                         "default": false
                     },
                     "azureFunctions.showProjectWarning": {
