+++ package.nls.json
@@ -22,7 +22,7 @@
     "azFunc.projectLanguageDescription": "The default language to use when performing operations in the Azure Functions extension (e.g. \"Create New Function\").",
     "azFunc.deploy": "Deploy to Function App...",
     "azFunc.configureDeploymentSource": "Configure Deployment Source...",
-    "azFunc.debugFunctionAppOnAzure": "Attach Debugger",
+    "azFunc.startJavaRemoteDebug": "Attach Debugger",
     "azFunc.appSettings.add": "Add New Setting...",
     "azFunc.appSettings.download": "Download Remote Settings...",
     "azFunc.appSettings.upload": "Upload Local Settings...",
@@ -46,7 +46,7 @@
     "azFunc.showDeploySubpathWarningDescription": "Show a warning when the \"deploySubpath\" setting does not match the selected folder for deploying.",
     "azFunc.startStreamingLogs": "Start Streaming Logs",
     "azFunc.stopStreamingLogs": "Stop Streaming Logs",
-    "azFunc.enableRemoteDebugging": "Enable remote debugging, an experimental feature that only supports Java-based Functions Apps.",
+    "azFunc.enableJavaRemoteDebugging": "Enable remote debugging, an experimental feature that only supports Java-based Functions Apps.",
     "azFunc.deleteProxy": "Delete Proxy...",
     "azFunc.pickProcessTimeoutDescription": "The timeout (in seconds) to be used when searching for the Azure Functions host process. Since a build is required every time you F5, you may need to adjust this based on how long your build takes.",
     "azFunc.templateVersion": "A runtime release version (any runtime) that species which templates will be used rather than the latest templates.  This version will be used for ALL runtimes. (Requires a restart of VS Code to take effect)",
